refactor and modal edits
